Split RLM composable env tests
